refactor: extract camera-ready config types
